Upload byte array to S3 to eliminate another intermediate local file
