+++ orchestration/temporal-java/src/main/java/edu/washu/tag/temporal/util/FileHandler.java
@@ -11,6 +11,9 @@ public interface FileHandler {
     @Retryable(maxAttempts = 5, backoff = @Backoff(delay = 1000, multiplier = 2))
     String putWithRetry(Path filePath, Path filePathsRoot, URI destination) throws IOException;
 
+    @Retryable(maxAttempts = 5, backoff = @Backoff(delay = 1000, multiplier = 2))
+    String putWithRetry(byte[] data, String relativePath, URI destination) throws IOException;
+
     String put(Path filePath, Path filePathsRoot, URI destination) throws IOException;
 
     List<String> put(List<Path> filePaths, Path filePathsRoot, URI destination) throws IOException;
